+++ tests/files/base.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
-USE_YAML = True
+USE_YAML = False
 TEST_PROP = 'BASE_PROPERTY'
 PREFIX = 'BASE_ENV'