+++ tests/test_get_settings.py
@@ -2,6 +2,7 @@
 import os
 
 import pytest
+from dotenv import load_dotenv
 
 from magic_settings.utils import BaseSettings, Property
 from tests.files import base, local
@@ -10,39 +11,85 @@ from tests.files import base, local
 TEST_DIR = os.path.dirname(os.path.abspath(__file__))
 
 
-@pytest.fixture(scope='module')
-def settings():
-    """Fixture settings"""
-    class TestSettings(BaseSettings):
-        USE_YAML = Property(types=bool)
-        TEST_PROP = Property(types=str)
-        PREFIX = Property(types=str)
-
-    settings = TestSettings()
-    return settings
+class TestSettings(BaseSettings):
+    USE_YAML = Property(types=bool, converts=[bool], default=False)
+    TEST_PROP = Property(types=str)
+    PREFIX = Property(types=str)
 
 
-def test_get_settings_with_params(settings):
-    """Test get_settings with all parameters"""
+@pytest.fixture
+def settings_with_yaml():
+    """Fixture settings with yaml"""
     dotenv_path = os.path.join(TEST_DIR, 'files', '.env')
     yaml_settings_path = os.path.join(TEST_DIR, 'files', 'settings.yaml')
-    settings.get_settings(
+    settings = TestSettings(
+        modules=[base, local],
         prefix='ENV',
         dotenv_path=dotenv_path,
         override_env=True,
         yaml_settings_path=yaml_settings_path,
-        use_yaml_settings=True,
-        base=base,
-        local=local,
     )
-    assert settings.USE_YAML is True
-    assert settings.PREFIX == 'YAML_ENV'
-    assert settings.TEST_PROP == 'YAML_PROPERTY'
+
+    return settings
 
 
-def test_get_settings_without_params(settings):
-    """Test get_settings without parameters"""
-    settings.get_settings(base=base, local=local)
-    assert settings.USE_YAML is True
-    assert settings.PREFIX == 'BASE_ENV'
-    assert settings.TEST_PROP == 'BASE_PROPERTY'
+@pytest.fixture
+def settings_with_env():
+    """Fixture settings with env"""
+    dotenv_path = os.path.join(TEST_DIR, 'files', '.env')
+    settings = TestSettings(modules=[base, local], prefix='ENV', dotenv_path=dotenv_path, override_env=True)
+    return settings
+
+
+@pytest.fixture
+def settings_with_prefix():
+    """Fixture settings with prefix"""
+    dotenv_path = os.path.join(TEST_DIR, 'files', '.env')
+    load_dotenv(dotenv_path=dotenv_path, override=True)
+    settings = TestSettings(prefix='ENV')
+    return settings
+
+
+@pytest.fixture
+def settings_with_modules():
+    """Fixture settings with modules"""
+    settings = TestSettings(modules=[base, local])
+    return settings
+
+
+def test_get_settings(settings_with_yaml):
+    """Test get_settings with yaml"""
+    settings_with_yaml.get_settings()
+    assert settings_with_yaml.USE_YAML
+    assert settings_with_yaml.PREFIX == 'YAML_ENV'
+    assert settings_with_yaml.TEST_PROP == 'YAML_PROPERTY'
+
+
+def test_get_settings_with_env(settings_with_env):
+    """Test get_settings with env"""
+    settings_with_env.get_settings()
+    assert settings_with_env.USE_YAML
+    assert settings_with_env.PREFIX == 'ENV_ENV'
+    assert settings_with_env.TEST_PROP == 'ENV_PROPERTY'
+
+
+def test_get_settings_with_prefix(settings_with_prefix):
+    """Test get_settings with prefix"""
+    settings_with_prefix.get_settings()
+    assert settings_with_prefix.USE_YAML
+    assert settings_with_prefix.PREFIX == 'ENV_ENV'
+    assert settings_with_prefix.TEST_PROP == 'ENV_PROPERTY'
+
+
+def test_get_settings_with_modules(settings_with_modules):
+    """Test get_settings with modules"""
+    settings_with_modules.get_settings()
+    assert not settings_with_modules.USE_YAML
+    assert settings_with_modules.PREFIX == 'BASE_ENV'
+    assert settings_with_modules.TEST_PROP == 'BASE_PROPERTY'
+
+
+def test_get_settings_with_bad_module():
+    """Test get_settings with bad parameters"""
+    with pytest.raises(ValueError, message='Expecting ValueError'):
+        TestSettings(modules=[base, local, 'not_module'])